Renamed pages.rb to base_page.rb; Rebased Users, Posts and Dashboard classes into single modules; Deleted requires from each single spec and rebased into spec_helper.rb.
