+++ spec/features/post_spec.rb
@@ -1,13 +1,10 @@
-require 'capybara/rspec'
 require_relative '../../spec/spec_helper'
-require_relative '../../pages'
-require_relative '../../locators'
 
 feature 'OpenSourceCMS Posts test', type: :feature do
   describe 'When on Posts Page' do
 
     before(:each) do
-      @posts_page = Pages::Posts.new
+      @posts_page = Posts.new
       @posts_page.open_main_page
     end
