+++ spec/spec_helper.rb
@@ -1,6 +1,10 @@
 require 'capybara'
 require 'selenium/webdriver'
-require_relative '../pages'
+require 'capybara/rspec'
+require_relative '../pages/base_page'
+require_relative '../pages/dashboard_page'
+require_relative '../pages/posts_page'
+require_relative '../pages/users_page'
 require_relative '../locators'
 
 Capybara.register_driver :selenium do |app|
